Defining observation schemes and loading data

Most often we collect multiple data points from a single observation scheme. Decorating each and single one of the observations as in the previous section might become quite tiring. To this end we provide a struct ObsScheme together with a function load_data for a convenient way of decorating multiple observations at once.

ObservationSchemes.ObsScheme — Type

struct ObsScheme{T}
    obs::T
    pattern::Vector{Int64}
    full_pattern::Vector{Int64}
    mode::Val
end

Defines an observation scheme for convenient loading of multiple data points at once and automatically equipping them with observation decorators stored in obs recycling through a set of indeces specified in pattern that pick the order of observation templates in obs that need to be applied.

ObservationSchemes.load_data — Function

load_data(os::ObsScheme, tt, xx)

Decorate the data in xx and times of recordings in tt according to an observation scheme template stored in os.

load_data(os::ObsScheme, tt_xx)

Same as load_data(os::ObsScheme, tt, xx), but tt_xx is a vector of tuples that pair the observation with its recorded time.

For instance, if all data is collected according to a single observation scheme:

t, v = 1.0, [1.0, 2.0, 3.0] # dummy values, only DataType matter
obs = LinearGsnObs(t, v; full_obs=true)

Then we can define it as an observation scheme and load the data as follows:

tt = [1.0, 2.0, 3.0, 4.0, 5.0]
xx = [i.+rand(3) for i in eachindex(tt)]

data = load_data(ObsScheme(obs), tt, xx)

If more than one type of observation schemes are used then we can specify a list of them to ObsScheme and then specify the pattern in which we cycle through them when loading the data:

# another observation scheme:
t, v = 1.0, SVector{2,Float64}(1.0, 2.0)
obs2 = LinearGsnObs(t, v; full_obs=true)

# observations:
xx = [i%2==1 ? i.+rand(3) : i.+rand(SVector{2,Float64}) for i in eachindex(tt)]

# load the data:
data = load_data(ObsScheme(obs, obs2; pattern=[1,2]), tt, xx)