Utility functions

We provide some additional utility functions.

ObservationSchemes.num_obs — Method

num_obs(all_obs::AllObservations)

Return the total number of observations in all recordings of all_obs.

ObservationSchemes.num_recordings — Method

num_recordings(all_obs::AllObservations)

Return the total number of recordings in all_obs.

ObservationSchemes.package — Function

package(objs::AbstractArray, all_obs::AllObservations)

package(objs::AbstractArray, recording::NamedTuple)

package(obj, recording::NamedTuple)

package(objs::AbstractArray, old_to_new_idx::Dict, all_obs::AllObservations)

package(obj, all_obs::AllObservations)

Various packaging functions that zip together objects/functions/structs together with observations defined in AllObservations/recording

ObservationSchemes.setup_time_grids — Function

setup_time_grids(
    all_obs::AllObservations,
    dt=0.01,
    τ=identity,
    eltype=Float64,
    already_arranged_tt=nothing
)

Same as a version for a single recording, but applies recursively to all recordings in all_obs.

ObservationSchemes.setup_time_grids — Function

setup_time_grids(
    recording::NamedTuple,
    dt=0.01,
    τ=identity,
    eltype=Float64,
    already_arranged_tt=nothing
)

Same as a version for a single interval, but applies recursively to all inter-observation intervals in recording.

ObservationSchemes.setup_time_grids — Function

setup_time_grids(
    t0::Number,
    T::Number,
    dt=0.01,
    τ=identity,
    eltype=Float64,
    already_arranged_tt=nothing
)

Set up a time grid for a single inter-observation interval [t0, T]. Starts with an equidistant time grid with mesh witdh dt and applies a transformation τ afterwards. Alternatively, if already_arranged_tt is passed then the initial setting up of an equidistant grid is omitted and already_arrange_tt is used in its place for the τ transformation.