Priors over starting points

The intention behind this package is to provide a complete infrastructure for defining observations for stochastic processes (though, it can find use for defining observations in other settings). Consequently, on top of decorating each observation with appropriate information, one often needs to specify some additional information regarding the starting point. This is because often no observations are made at the initial time, but we still wish to conveys our prior knowledge of where the point lies via prior distributions. For this reason the starting points are always treated differently to observations and this package provides an infrastructure for specifying priors over starting points.

All priors over starting points inherit from

ObservationSchemes.StartingPtPrior — Type

StartingPtPrior

Types inheriting from the abstract type StartingPtPrior indicate the prior that is put on the starting point of the observed path of some stochastic process

They all must implement methods rand, start_pt, and logpdf (and should also implement inv_start_pt for MCMC setting). In this package we provide implementations for the following types of starting points

Known, fixed starting points

Gaussian priors over starting points

Known starting point

This is the simplest setting in which the starting point is assumed to be known.

ObservationSchemes.KnownStartingPt — Type

struct KnownStartingPt{T} <: StartingPtPrior{T}
    y::T
end

Indicates that the starting point is known and stores its value in y

It can be defined with

x0 = [1.0, 2.0]
x0_prior = KnownStartingPt(x0)

a call to rand or start_pt will simply return the fixed starting point and logpdf(x0_prior, y) evaluates to 0 so long as x0 == y.

Gaussian priors

A Gaussian prior over the starting point.

ObservationSchemes.GsnStartingPt — Type

struct GsnStartingPt{T,S,TM} <: StartingPtPrior{T} where {S}
    μ::T
    Σ::S
    Λ::S
    μ₀::T
    Σ₀::UniformScaling
end

Indicates that the starting point is equipped with a Gaussian prior with mean μ and covariance matrix Σ (and pre-computed precision Λ:=Σ⁻¹). Sampling is always done via non-centred parametrization, by sampling white noise z according to Gaussian with zero mean and identity covariance: μ₀ and Σ₀, and then transforming z to a variable with mean and covariance μ and Σ.

Can be defined with

μ, Σ = [1.0, 2.0], [1.0 0.0; 0.0 1.0]
x0_prior = GsnStartingPt(μ, Σ)

to set the mean and covariance to μ and Σ respectively. The underlying idea behind Gaussian starting point priors is that of non-centred parametrisation, so that a possibility of local updates is granted. More precisely any sampling is done with z∼N(0,Id) variables, which are then transformed to N(μ,Σ) via linear transformations. In particular, sampling with rand can be done with local perturbations via Crank-Nicolson scheme.

Base.rand — Function

rand([rng::Random.AbstractRNG], G::GsnStartingPt, z, ρ=0.0)

Sample new white noise using Crank-Nicolson scheme with memory parameter ρ and a previous value of the white noise stored inside object G

rand([rng::Random.AbstractRNG], G::GsnStartingPt)

Sample new starting point according to its prior distribution.

rand([rng::Random.AbstractRNG], G::KnownStartingPt, args...)

Starting point is known. Nothing can be sampled. Returning known starting point.

inv_start_pt returns the non-centrally parametrised noise z that produces a given starting point x0:

ObservationSchemes.inv_start_pt — Function

inv_start_pt(y, G::GsnStartingPt, P)

Compute the driving noise that is needed to obtain starting point y under prior G and the likelihood in P

inv_start_pt(y, G::KnownStartingPt, P)

Starting point known, no need for dealing with white noise, use convention of returning y

and start_pt is the reverse operation

ObservationSchemes.start_pt — Function

start_pt(z, G::GsnStartingPt, P)

Compute a new starting point from the white noise for a given posterior distribution obtained from combining prior G and the likelihood encoded by the object P.

start_pt(z, G::GsnStartingPt)

Compute a new starting point from the white noise for a given prior distribution G

start_pt(z, G::KnownStartingPt, P)

Return starting point

start_pt(G::KnownStartingPt, P)

Return starting point

Tip

To see how to define your own priors over starting points see a How-to-guide on Defining custom priors over starting points