Utility functions

We provide some additional utility functions.

ObservationSchemes.build_recording — Method

build_recording(
        ::Type{K}, tt, observs::Vector, P, t0, x0_prior; kwargs...
    ) where K

Utility function to build a recording that follows the convention adopted in package ObservationSchemes.

ObservationSchemes.num_obs — Method

num_obs(all_obs::AllObservations)

Return total number of observations

ObservationSchemes.num_recordings — Method

num_recordings(all_obs::AllObservations)

Return total number of recordings

ObservationSchemes.package — Function

package

Various packaging functions that zip together objects/functions/structs together with observations defined in AllObservations/recording

ObservationSchemes.setup_time_grids — Function

setup_time_grids(
    all_obs::AllObservations,
    dt=0.01,
    τ=identity,
    eltype=Float64,
    already_arranged_tt=nothing
)

Same as a version for a single recording, but applies recursively to all recordings in all_obs.

ObservationSchemes.setup_time_grids — Function

setup_time_grids(
    t0::Number,
    T::Number,
    dt=0.01,
    τ=identity,
    eltype=Float64,
    already_arranged_tt=nothing
)

Set up a time grid for a single inter-observation interval [t0, T]. Starts with an equidistant time grid with mesh witdh dt and applies a transformation τ afterwards. Alternatively, if already_arranged_tt is passed then the initial setting up of an equidistant grid is omitted and already_arrange_tt is used in its place for the τ transformation.

ObservationSchemes.setup_time_grids — Function

setup_time_grids(
    recording::NamedTuple,
    dt=0.01,
    τ=identity,
    eltype=Float64,
    already_arranged_tt=nothing
)

Set up a time grid for a single recording. Starts with an equidistant time grid with mesh witdh dt and applies a transformation τ to each inter-observation integral. Alternatively, if already_arranged_tt is passed then the initial setting up of equidistant grid is omitted and already_arrange_tt is used in its place for the τ transformation.