ObservationSchemes.jl

ObservationSchemes.AllObservations

ObservationSchemes.FirstPassageAbstract

ObservationSchemes.FirstPassageTimeInfo

ObservationSchemes.GsnStartingPt

ObservationSchemes.KnownStartingPt

ObservationSchemes.LinearGsnObs

ObservationSchemes.NoFirstPassageTimes

ObservationSchemes.Observation

ObservationSchemes.StartingPtPrior

Base.rand

Base.rand

Distributions.logpdf

Distributions.logpdf

ObservationSchemes.add_dependency!

ObservationSchemes.add_recording!

ObservationSchemes.check_recording_format

ObservationSchemes.fill_dependency_for_unspec_params!

ObservationSchemes.format_dep_entry

ObservationSchemes.format_law_dep_entry

ObservationSchemes.format_law_dep_entry

ObservationSchemes.format_law_dep_entry

ObservationSchemes.fpt_info

ObservationSchemes.initialize!

ObservationSchemes.inv_start_pt

ObservationSchemes.inv_start_pt

ObservationSchemes.new_param_dependence!

ObservationSchemes.parameter_names

ObservationSchemes.split_recording!

ObservationSchemes.start_pt

ObservationSchemes.start_pt

ObservationSchemes.start_pt

ObservationSchemes.start_pt

ObservationSchemes.AllObservations — Type

AllObservations

A struct gathering multiple observations of a diffusion processes. Additionaly, the interdependence structure between parameters shared between various diffusions laws used to generate the recorded data is kept.

    AllObservations(;P=nothing, obs=nothing)

Default constructor creating either an empty `AllObservations` object,
or initiates it immediately with a single recording where the target comes
from the law `P`, the observations are stored in `obs` and the observed
process was started at time `t0` from som position which we put a prior on
`x0_prior`.

    AllObservations(recording::NamedTuple)

Constructor creating an `AllObservations` object and initiating it
immediately with a single recording where the target comes
from the law `recording.P`, the observations are stored in `recording.obs`
and the starting point is at time `recording.t0` and has a prior `x0_prior`.

ObservationSchemes.FirstPassageTimeInfo — Type

FirstPassageTimeInfo{C,L,U,A} <: FirstPassageAbstract

Compile-time indicator for the first-passage time information conveyed by the data-point. C lists the affected coordinates (of the observations, not the process), L indicates the level, U are indicators for whether the respective first-passage times are up-crossings and A is and indicator for whether an additional reset needs to be reached before first-passage time (note that reset time is defined as a first-passage time to a corresponding reset level stored in R that happens anytime before the actual first-passage time, the reset-time happens in the direction opposite to the direction of first-passage time crossing; additionally, the coordinate can reach the first-passage time level prior to being reset).

ObservationSchemes.GsnStartingPt — Type

GsnStartingPt{T,S} <: StartingPtPrior

Indicates that the starting point is equipped with a Gaussian prior with mean μ and covariance matrix Σ. It also stores the most recently sampled white noise z used to compute the starting point and a precision matrix Λ:=Σ⁻¹. μ₀ and Σ₀ are the mean and covariance of the white noise

GsnStartingPt(μ::T, Σ::S)

Base constructor. It initialises the mean μ and covariance Σ parameters and Λ is set according to Λ:=Σ⁻¹

ObservationSchemes.KnownStartingPt — Type

KnownStartingPt{T} <: StartingPtPrior

Indicates that the starting point is known and stores its value in y

ObservationSchemes.LinearGsnObs — Type

LinearGsnObs{D,T,FPT,S,R} <: Observation{D,T}

Observation of the underlying process that is of the form: LX+ξ, where ξ∼N(μ,Σ) and L,Σ and μ are respectively matrices and a vector of appropriate dimensions. FPT stores information about first-passage times. full_obs is an indicator for whether it is a full observation of the process (as it grants the use of Markov Property)

ObservationSchemes.NoFirstPassageTimes — Type

NoFirstPassageTimes <: FirstPassageAbstract

Compile-time indicator that the observation stores no first-passage time data

ObservationSchemes.add_dependency! — Method

add_dependency!(all_obs::AllObservations, dep::Dict)

Add a dependency structure dep between parameters shared across various laws and observations used to generate various recordings stored in an observations container all_obs.

ObservationSchemes.add_recording! — Method

add_recording!(all_obs::AllObservations, recording::NamedTuple)

Add a new recording recording to observations container all_obs.

ObservationSchemes.fpt_info — Method

fpt_info(::LinearGsnObs{D,T,FPT}) where {D,T,FPT}

Return information about first-passage times

ObservationSchemes.initialize! — Method

initialize(all_obs::AllObservations)

Split the recordings at the times of full observations to make full use of the Markov property (and make the code readily parallelisable).

ObservationSchemes.FirstPassageAbstract — Type

FirstPassageAbstract

Types ihneriting from this struct define the type of information stored about first-passage times

ObservationSchemes.Observation — Type

Observation{D,T}

Types inheriting from this struct provide all information about the observation of the stochastic process being made. D denotes the dimension of the observation whereas T its datatype.

ObservationSchemes.StartingPtPrior — Type

StartingPtPrior

Types inheriting from the abstract type StartingPtPrior indicate the prior that is put on the starting point of the observed path following the dynamics of some stochastic differential equation.

Base.rand — Function

rand(G::KnownStartingPt, ::Any)

If starting point is known then nothing can be sampled, using convention of returning the known startin point

Base.rand — Method

rand(G::GsnStartingPt, ρ)

Sample new white noise using Crank-Nicolson scheme with memory parameter ρ and a previous value of the white noise stored inside object G

Distributions.logpdf — Method

logpdf(::GsnStartingPt, y)

log-probability density function evaluated at y of a prior distribution G

Distributions.logpdf — Method

logpdf(::KnownStartingPt, y)

Nothing to do so long as y is equal to the known starting point inside G

ObservationSchemes.check_recording_format — Method

check_recording_format(::T) where T<:NamedTuple

Check whether the namedtuple has fields P, obs, t0 and x0_prior that each recording must have.

ObservationSchemes.fill_dependency_for_unspec_params! — Method

fill_dependency_for_unspec_params!(all_obs::AllObservations)

Fill the all_obs.depen_param dictionary with all parameters that are not shared between recordings.

ObservationSchemes.format_dep_entry — Method

format_dep_entry(P, entry)

Format an entry pointing to a specific parameter in a specific law used for a specific recording so that it encodes recording index rec_idx, parameter index param_idx and parameter name param_name of the pointing to position.

ObservationSchemes.format_law_dep_entry — Method

format_dep_entry(P, entry::Tuple{Int64,Int64,Symbol})

Make sure that recording index rec_idx, parameter index param_idx and parameter name param_name make sense and are formatted correctly.

ObservationSchemes.format_law_dep_entry — Method

format_dep_entry(P, entry::Tuple{Int64,Int64})

Infer recording index rec_idx, parameter index param_idx and parameter name param_name from the pair that lists recording index and parameter index within the law that was used for that recording.

ObservationSchemes.format_law_dep_entry — Method

format_dep_entry(P, entry::Tuple{Int64,Symbol})

Infer recording index rec_idx, parameter index param_idx and parameter name param_name from the pair that lists recording index and parameter name within the law that was used for that recording.

ObservationSchemes.inv_start_pt — Method

inv_start_pt(y, G::GsnStartingPt, P)

Compute the driving noise that is needed to obtain starting point y under prior G and the likelihood in P

ObservationSchemes.inv_start_pt — Method

inv_start_pt(y, G::KnownStartingPt, P)

Starting point known, no need for dealing with white noise, use convention of returning y

ObservationSchemes.new_param_dependence! — Method

new_param_dependence!(out, old_dep, old_to_new_idx)

Define a new parameter dependence structure by using an old parameter dependence saved in old_dep and deducing the changed indices of affected recording by following the old_to_new_idx dictionary. The new dependency is saved directly to out structure.

ObservationSchemes.parameter_names — Function

parameter_names

Return names of the parameters of a given object

ObservationSchemes.split_recording! — Method

split_recording!(
        recording_idx,
        recording,
        counter,
        out::AllObservations,
        old_to_new_idx,
        old_to_new_obs_idx,
    )

Split a single record recording at the times of full observations. Save new, splitted recordings in the struct out. Keep account of the changed indices of the recordings for the recording index <---> parameter dependence pairs by updating a suitable dictionary old_to_new_idx, starting the new recording's index at an offset counter.

ObservationSchemes.start_pt — Method

start_pt(G::GsnStartingPt, P)

Compute a new starting point from the white noise for a given posterior distribution obtained from combining prior G and the likelihood encoded by the object P.

ObservationSchemes.start_pt — Method

start_pt(G::GsnStartingPt, P)

Compute a new starting point from the white noise for a given prior distribution G

ObservationSchemes.start_pt — Method

start_pt(G::KnownStartingPt, P)

Return starting point

ObservationSchemes.start_pt — Method

start_pt(G::KnownStartingPt, P)

Return starting point