ObservationSchemes.jl

ObservationSchemes.AllObservations

ObservationSchemes.FirstPassageAbstract

ObservationSchemes.FirstPassageTimeInfo

ObservationSchemes.GeneralObs

ObservationSchemes.GsnStartingPt

ObservationSchemes.GsnStartingPt

ObservationSchemes.KnownStartingPt

ObservationSchemes.LinearGsnObs

ObservationSchemes.NoFirstPassageTimes

ObservationSchemes.ObsScheme

ObservationSchemes.Observation

ObservationSchemes.ParametrizedObservation

ObservationSchemes.StartingPtPrior

Base.eltype

Base.length

Base.rand

Base.rand

Base.rand

Base.rand

Base.size

Distributions.logpdf

Distributions.logpdf

ObservationSchemes.L

ObservationSchemes.L

ObservationSchemes.add_dependency!

ObservationSchemes.add_dependency!

ObservationSchemes.add_recording!

ObservationSchemes.build_recording

ObservationSchemes.check_recording_format

ObservationSchemes.fill_dependency_for_unspec_params!

ObservationSchemes.format_dep_entry

ObservationSchemes.format_law_dep_entry

ObservationSchemes.format_law_dep_entry

ObservationSchemes.format_law_dep_entry

ObservationSchemes.fpt_info

ObservationSchemes.initialize!

ObservationSchemes.inv_start_pt

ObservationSchemes.inv_start_pt

ObservationSchemes.inv_start_pt

ObservationSchemes.load_data

ObservationSchemes.load_data

ObservationSchemes.load_data

ObservationSchemes.new_param_dependence!

ObservationSchemes.num_obs

ObservationSchemes.num_recordings

ObservationSchemes.obs

ObservationSchemes.package

ObservationSchemes.parameter_names

ObservationSchemes.set_parameters!

ObservationSchemes.setup_time_grids

ObservationSchemes.setup_time_grids

ObservationSchemes.setup_time_grids

ObservationSchemes.split_recording!

ObservationSchemes.start_pt

ObservationSchemes.start_pt

ObservationSchemes.start_pt

ObservationSchemes.start_pt

ObservationSchemes.start_pt

ObservationSchemes.Λ

ObservationSchemes.Σ

ObservationSchemes.Σ

ObservationSchemes.μ

ObservationSchemes.μ

ObservationSchemes.ν

ObservationSchemes.ν

ObservationSchemes.FirstPassageTimeInfo — Type

FirstPassageTimeInfo{C,L,U,A} <: FirstPassageAbstract

Compile-time indicator for the first-passage time information conveyed by the data-point. C lists the affected coordinates (of the observations, not the process), L indicates the level, U are indicators for whether the respective first-passage times are up-crossings and A is and indicator for whether an additional reset needs to be reached before first-passage time (note that reset time is defined as a first-passage time to a corresponding reset level stored in R that happens anytime before the actual first-passage time, the reset-time happens in the direction opposite to the direction of first-passage time crossing; additionally, the coordinate can reach the first-passage time level prior to being reset).

ObservationSchemes.GsnStartingPt — Method

GsnStartingPt(μ::T, Σ::S)

Base constructor of GsnStartingPt. It initialises the mean μ and covariance Σ parameters and Λ is set according to Λ:=Σ⁻¹

ObservationSchemes.NoFirstPassageTimes — Type

NoFirstPassageTimes <: FirstPassageAbstract

Compile-time indicator that the observation stores no first-passage time data

ObservationSchemes.add_dependency! — Method

add_dependency!(all_obs::AllObservations, dep::Dict)

Add a dependency structure dep between parameters shared across various laws and observations used to generate various recordings stored in an observations container all_obs.

ObservationSchemes.add_recording! — Method

add_recording!(all_obs::AllObservations, recording::NamedTuple)

Add a new recording recording to observations container all_obs.

ObservationSchemes.fpt_info — Method

fpt_info(::LinearGsnObs{D,T,Tag,FPT}) where {D,T,Tag,FPT}

Return information about first-passage times

ObservationSchemes.initialize! — Method

initialize!(all_obs::AllObservations)

Split the recordings at the times of full observations to make full use of the Markov property (and make the code readily parallelisable).

ObservationSchemes.load_data — Method

load_data(os::ObsScheme, tt, xx)

Decorate the data in xx and times of recordings in tt according to an observation scheme template stored in os.

ObservationSchemes.load_data — Method

load_data(os::ObsScheme, tt_xx)

Same as load_data(os::ObsScheme, tt, xx), but tt_xx is a vector of tuples that pair the observation with its recorded time.

ObservationSchemes.FirstPassageAbstract — Type

FirstPassageAbstract

Types ihneriting from this struct define the type of information stored about first-passage times

ObservationSchemes.Observation — Type

Observation{D,T}

Types inheriting from this struct provide all information about the observation of the stochastic process being made. D denotes the dimension of the observation whereas T its eltype.

ObservationSchemes.ParametrizedObservation — Type

ParametrizedObservation{D,T} <: Observation{D,T}

Same as observations but additionally, types inheriting from this struct depend on some parameters (that may in practice be unknown).

Base.eltype — Method

eltype(::K) where {K<:Observation{D,T}}

Eltype of the observation.

Base.length — Method

length(::K) where {K<:Observation{D}}

Length of the observation (equal to a number of entries in a vector holding the data)

Base.rand — Function

rand([rng::Random.AbstractRNG], G::GsnStartingPt, z, ρ=0.0)

Sample new white noise using Crank-Nicolson scheme with memory parameter ρ and a previous value of the white noise stored inside object G

Base.rand — Method

rand([rng::Random.AbstractRNG], G::GsnStartingPt)

Sample new starting point according to its prior distribution.

Base.rand — Method

rand([rng::Random.AbstractRNG], G::KnownStartingPt, args...)

Starting point is known. Nothing can be sampled. Returning known starting point.

Base.size — Method

size(::K) where {K<:Observation{D}}

Size of the observation

Distributions.logpdf — Method

logpdf(::GsnStartingPt, y)

log-probability density function evaluated at y of a prior distribution G

Distributions.logpdf — Method

logpdf(::KnownStartingPt, y)

Nothing to do so long as y is equal to the known starting point inside G

ObservationSchemes.L — Method

L(o::LinearGsnObs)

Return matrix L from the observation scheme ν = Lx+ξ, where ξ∼N(μ,Σ)

ObservationSchemes.check_recording_format — Method

check_recording_format(::T) where T<:NamedTuple

Check whether the namedtuple has fields P, obs, t0 and x0_prior that each recording must have.

ObservationSchemes.fill_dependency_for_unspec_params! — Method

fill_dependency_for_unspec_params!(all_obs::AllObservations)

Fill the all_obs.depen_param dictionary with all parameters that are not shared between recordings.

ObservationSchemes.format_dep_entry — Method

format_dep_entry(P, entry)

Format an entry pointing to a specific parameter in a specific law used for a specific recording so that it encodes recording index rec_idx, parameter index param_idx and parameter name param_name of the pointing to position.

ObservationSchemes.format_law_dep_entry — Method

format_dep_entry(P, entry::Tuple{Int64,Int64,Symbol})

Make sure that recording index rec_idx, parameter index param_idx and parameter name param_name make sense and are formatted correctly.

ObservationSchemes.format_law_dep_entry — Method

format_dep_entry(P, entry::Tuple{Int64,Int64})

Infer recording index rec_idx, parameter index param_idx and parameter name param_name from the pair that lists recording index and parameter index within the law that was used for that recording.

ObservationSchemes.format_law_dep_entry — Method

format_dep_entry(P, entry::Tuple{Int64,Symbol})

Infer recording index rec_idx, parameter index param_idx and parameter name param_name from the pair that lists recording index and parameter name within the law that was used for that recording.

ObservationSchemes.inv_start_pt — Method

inv_start_pt(y, G::GsnStartingPt, P)

Compute the driving noise that is needed to obtain starting point y under prior G and the likelihood in P

ObservationSchemes.inv_start_pt — Method

inv_start_pt(y, G::KnownStartingPt, P)

Starting point known, no need for dealing with white noise, use convention of returning y

ObservationSchemes.new_param_dependence! — Method

new_param_dependence!(out, old_dep, old_to_new_idx)

Define a new parameter dependence structure by using an old parameter dependence saved in old_dep and deducing the changed indices of affected recording by following the old_to_new_idx dictionary. The new dependency is saved directly to out structure.

ObservationSchemes.parameter_names — Function

parameter_names

Return names of the parameters of a given object

ObservationSchemes.set_parameters! — Method

set_parameters!(objects, all_obs::AllObservations, coords, θ°)

Set new parameters in objects. New parameters are listed in a vector θ°. The coords lists the subset of coordinates from the total vector 𝛩 that θ° is supposed to replace. all_obs allows to decipher which stores which object in objects depends on any subset of parameters. objects is a list of objects in shape corresponding to param_depend whose parameters are to be updated.

ObservationSchemes.split_recording! — Method

split_recording!(
        recording_idx,
        recording,
        counter,
        out::AllObservations,
        old_to_new_idx,
        old_to_new_obs_idx,
    )

Split a single record recording at the times of full observations. Save new, splitted recordings in the struct out. Keep account of the changed indices of the recordings for the recording index <---> parameter dependence pairs by updating a suitable dictionary old_to_new_idx, starting the new recording's index at an offset counter.

ObservationSchemes.start_pt — Method

start_pt(z, G::GsnStartingPt, P)

Compute a new starting point from the white noise for a given posterior distribution obtained from combining prior G and the likelihood encoded by the object P.

ObservationSchemes.start_pt — Method

start_pt(z, G::GsnStartingPt)

Compute a new starting point from the white noise for a given prior distribution G

ObservationSchemes.start_pt — Method

start_pt(z, G::KnownStartingPt, P)

Return starting point

ObservationSchemes.start_pt — Method

start_pt(G::KnownStartingPt, P)

Return starting point

ObservationSchemes.Λ — Method

Λ(o::LinearGsnObs)

Return matrix Λ:=Σ⁻¹ from the observation scheme ν = Lx+ξ, where ξ∼N(μ,Σ)

ObservationSchemes.Σ — Method

Σ(o::LinearGsnObs)

Return matrix Σ from the observation scheme ν = Lx+ξ, where ξ∼N(μ,Σ)

ObservationSchemes.μ — Method

μ(o::LinearGsnObs)

Return vector μ from the observation scheme ν = Lx+ξ, where ξ∼N(μ,Σ)

ObservationSchemes.ν — Method

ν(o::Observation)

Return the observation